# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Austin Rifle Club API",
    "description": "API for Austin Rifle Club member portal, range status, events, and administration.",
    "version": "1.0.0",
    "contact": {
      "email": "info@austinrifleclub.org"
    }
  },
  "servers": [
    {
      "url": "https://austinrifleclub.org/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:8787/api",
      "description": "Development"
    }
  ],
...
  "paths": {
...
    "/applications": {
      "get": {
        "tags": ["Applications"],
        "summary": "List applications (admin only)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Application list" }
        }
      },
      "post": {
        "tags": ["Applications"],
        "summary": "Submit membership application",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApplication" }
            }
          }
        },
        "responses": {
          "201": { "description": "Application submitted" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "CreateApplication": {
        "type": "object",
        "required": ["firstName", "lastName", "email", "phone"],
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "address": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "zip": { "type": "string" },
          "membershipType": { "type": "string" },
          "referredBy": { "type": "string" },
          "shootingInterests": { "type": "array", "items": { "type": "string" } }
        }
      },
...
    }
  }
}